Update index.html
Removed the script tag from bottom to top, adding "defer" to the tag would still work the same way as expected, adding a script src tag to the bottom of the HTML document is an old way of doing things, now "defer" is here to help. 

Also filled in the gaps for your alt attributes, just in case.